+++ src/igraph/io/adjacency.py
