+++ src/igraph/io/bipartite.py
