+++ src/igraph/io/random.py
